+++ Modules/Microsoft365DSC/Examples/Resources/EXOOfflineAddressBook/2-Update.ps1
@@ -20,8 +20,8 @@ Configuration Example
         {
             Name                 = "Integration Address Book"
             AddressLists         = @('\All Users')
-            DiffRetentionPeriod  = "30"
-            IsDefault            = $false # Updated Property
+            DiffRetentionPeriod  = "60" # Updated Property
+            IsDefault            = $true
             Ensure               = "Present"
             Credential           = $Credscredential
         }