+++ Modules/Microsoft365DSC/Modules/M365DSCTelemetryEngine.psm1
@@ -229,7 +229,7 @@ function Add-M365DSCTelemetryEvent
                 $Data.Add('LCMState', $LCMInfo.LCMState)
                 $Data.Add('LCMStateDetail', $LCMInfo.LCMStateDetail)
 
-                if ([System.String]::IsNullOrEMpty($Type))
+                if ([System.String]::IsNullOrEmpty($Type))
                 {
                     if ($Global:M365DSCExportInProgress)
                     {
@@ -263,6 +263,14 @@ function Add-M365DSCTelemetryEvent
 
             $M365DSCTelemetryEventId = (New-GUID).ToString()
             $Data.Add('M365DSCTelemetryEventId', $M365DSCTelemetryEventId)
+
+            if ([System.String]::IsNullOrEMpty($Type))
+            {
+                if ((-not [System.String]::IsNullOrEmpty($Data.Method) -and $Data.Method -eq 'Export-TargetResource') -or $Global:M365DSCExportInProgress)
+                {
+                    $Type = 'Export'
+                }
+            }
 
             $TelemetryClient.TrackEvent($Type, $Data, $Metrics)
         }