+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -514,7 +514,6 @@ function Get-M365DSCTenantNameFromParameterSet
         [System.Collections.HashTable]
         $ParameterSet
     )
-
     if ($ParameterSet.TenantId)
     {
         return $ParameterSet.TenantId
@@ -578,7 +577,12 @@ function Test-M365DSCParameterState
     #endregion
     $returnValue = $true
 
-    $DriftedParameters = @{ }
+    $DriftedParameters = @{}
+    $DriftObject = @{
+        DriftInfo     = @{}
+        CurrentValues = @{}
+        DesiredValues = @{}
+    }
     if ($null -ne $IncludedDrifts -and $IncludedDrifts.Keys.Count -gt 0)
     {
         $DriftedParameters = $IncludedDrifts
@@ -644,6 +648,7 @@ function Test-M365DSCParameterState
                                     'values, but it was either not present or ' + `
                                     'was null. This has caused the test method ' + `
                                     'to return false.')
+                            $DriftObject.DriftInfo.Add($fieldName, '')
                             $DriftedParameters.Add($fieldName, '')
                             $returnValue = $false
                         }
@@ -680,6 +685,11 @@ function Test-M365DSCParameterState
 
                                     if (-not $DriftedParameters.ContainsKey($fieldName))
                                     {
+                                        $DriftObject.DriftInfo.Add($fieldName, @{
+                                            PropertyName = $item.PropertyName
+                                            CurrentValue = $item.CurrentValue
+                                            DesiredValue = $item.DesiredValue
+                                        })
                                         $DriftedParameters.Add($fieldName, $EventValue)
                                     }
                                 }
@@ -703,6 +713,10 @@ function Test-M365DSCParameterState
 
                                 $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                 $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                $DriftObject.DriftInfo.Add($fieldName, @{
+                                    CurrentValue = $CurrentValues.$fieldName
+                                    DesiredValue = $DesiredValues.$fieldName
+                                })
                                 $DriftedParameters.Add($fieldName, $EventValue)
                                 $returnValue = $false
                             }
@@ -728,6 +742,10 @@ function Test-M365DSCParameterState
                                             "'$($DesiredValues.$fieldName)'")
                                     $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                     $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                    $DriftObject.DriftInfo.Add($fieldName, @{
+                                        CurrentValue = $CurrentValues.$fieldName
+                                        DesiredValue = $DesiredValues.$fieldName
+                                    })
                                     $DriftedParameters.Add($fieldName, $EventValue)
                                     $returnValue = $false
                                 }
@@ -748,6 +766,10 @@ function Test-M365DSCParameterState
                                             "'$($DesiredValues.$fieldName)'")
                                     $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                     $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                    $DriftObject.DriftInfo.Add($fieldName, @{
+                                        CurrentValue = $CurrentValues.$fieldName
+                                        DesiredValue = $DesiredValues.$fieldName
+                                    })
                                     $DriftedParameters.Add($fieldName, $EventValue)
                                     $returnValue = $false
                                 }
@@ -768,6 +790,10 @@ function Test-M365DSCParameterState
                                             "'$($DesiredValues.$fieldName)'")
                                     $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                     $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                    $DriftObject.DriftInfo.Add($fieldName, @{
+                                        CurrentValue = $CurrentValues.$fieldName
+                                        DesiredValue = $DesiredValues.$fieldName
+                                    })
                                     $DriftedParameters.Add($fieldName, $EventValue)
                                     $returnValue = $false
                                 }
@@ -784,6 +810,10 @@ function Test-M365DSCParameterState
                                             "'$($DesiredValues.$fieldName)'")
                                     $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                     $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                    $DriftObject.DriftInfo.Add($fieldName, @{
+                                        CurrentValue = $CurrentValues.$fieldName
+                                        DesiredValue = $DesiredValues.$fieldName
+                                    })
                                     $DriftedParameters.Add($fieldName, $EventValue)
                                     $returnValue = $false
                                 }
@@ -804,6 +834,10 @@ function Test-M365DSCParameterState
                                             "'$($DesiredValues.$fieldName)'")
                                     $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                     $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+                                    $DriftObject.DriftInfo.Add($fieldName, @{
+                                        CurrentValue = $CurrentValues.$fieldName
+                                        DesiredValue = $DesiredValues.$fieldName
+                                    })
                                     $DriftedParameters.Add($fieldName, $EventValue)
                                     $returnValue = $false
                                 }
@@ -859,6 +893,11 @@ function Test-M365DSCParameterState
                                             if (-not $DriftedParameters.ContainsKey($fieldName))
                                             {
                                                 $DriftedParameters.Add($fieldName, $EventValue)
+                                                $DriftObject.DriftInfo.Add($fieldName, @{
+                                                    PropertyName = $item.PropertyName
+                                                    CurrentValue = $item.CurrentValue
+                                                    DesiredValue = $item.DesiredValue
+                                                })
                                             }
                                         }
                                         $returnValue = $false
@@ -873,6 +912,11 @@ function Test-M365DSCParameterState
                                         'Test-M365DSCParameterState cmdlet')
                                 $EventValue = "<CurrentValue>$($CurrentValues.$fieldName)</CurrentValue>"
                                 $EventValue += "<DesiredValue>$($DesiredValues.$fieldName)</DesiredValue>"
+
+                                $DriftObject.DriftInfo.Add($fieldName, @{
+                                    CurrentValue = $CurrentValues.$fieldName
+                                    DesiredValue = $DesiredValues.$fieldName
+                                })
                                 $DriftedParameters.Add($fieldName, $EventValue)
                                 $returnValue = $false
                             }
@@ -898,47 +942,37 @@ function Test-M365DSCParameterState
         $EventMessage = [System.Text.StringBuilder]::New()
         $EventMessage.Append("<M365DSCEvent>`r`n") | Out-Null
         $TenantName = Get-M365DSCTenantNameFromParameterSet -ParameterSet $DesiredValues
+        Write-Verbose -Message "Found Tenant Name: $TenantName"
         $EventMessage.Append("    <ConfigurationDrift Source=`"$Source`" TenantId=`"$TenantName`">`r`n") | Out-Null
-
         $EventMessage.Append("        <ParametersNotInDesiredState>`r`n") | Out-Null
+
+        $driftedData = [System.Collections.Generic.Dictionary[[String], [String]]]::new()
+        $driftedData.Add('Tenant', $TenantName)
+        $DriftObject.Add('Tenant', $TenantName)
+        $driftedData.Add('Resource', $source.Split('_')[1])
+        $DriftObject.Add('Resource', $source.Split('_')[1])
+        $driftedData.Add('Event', 'DriftedParameter')
+
+        # If custom App Insights is specified, allow for the current and desired values to be captured;
+        # ISSUE #1222
+        if ($null -ne $env:M365DSCTelemetryInstrumentationKey -and `
+                $env:M365DSCTelemetryInstrumentationKey -ne 'bc5aa204-0b1e-4499-a955-d6a639bdb4fa' -and `
+                $env:M365DSCTelemetryInstrumentationKey -ne 'e670af5d-fd30-4407-a796-8ad30491ea7a')
+        {
+            $driftedData.Add('CurrentValues', $CurrentValues)
+            $driftedData.Add('DesiredValues', $DesiredValues)
+        }
+        #endregion
+        $telemetryDriftedParameters = ''
         foreach ($key in $DriftedParameters.Keys)
         {
             Write-Verbose -Message "Detected Drifted Parameter [$Source]$key"
-
-            #region Telemetry
-            $driftedData = [System.Collections.Generic.Dictionary[[String], [String]]]::new()
-            $driftedData.Add('Event', 'DriftedParameter')
-            $driftedData.Add('Parameter', "[$Source]$key")
-
-            # If custom App Insights is specified, allow for the current and desired values to be captured;
-            # ISSUE #1222
-            if ($null -ne $env:M365DSCTelemetryInstrumentationKey -and `
-                    $env:M365DSCTelemetryInstrumentationKey -ne 'bc5aa204-0b1e-4499-a955-d6a639bdb4fa' -and `
-                    $env:M365DSCTelemetryInstrumentationKey -ne 'e670af5d-fd30-4407-a796-8ad30491ea7a')
-            {
-                $driftedData.Add('CurrentValue', [string]($CurrentValues[$key]))
-                $driftedData.Add('DesiredValue', [string]($DesiredValues[$key]))
-            }
-            if (-not $Data.ContainsKey('Tenant'))
-            {
-                $driftedData.Add('Tenant', $TenantName)
-            }
-            $driftedData.Add('Resource', $source.Split('_')[1])
-            Add-M365DSCTelemetryEvent -Type 'DriftInfo' -Data $driftedData
-            #endregion
+            $telemetryDriftedParameters += $key + "`r`n"
             $EventMessage.Append("            <Param Name=`"$key`">" + $DriftedParameters.$key + "</Param>`r`n") | Out-Null
         }
 
-        #region Telemetry
-        $TenantName = Get-M365DSCTenantNameFromParameterSet -ParameterSet $DesiredValues
-        $data.Add('Event', 'ConfigurationDrift')
-
-        if (-not $Data.ContainsKey('Tenant'))
-        {
-            $data.Add('Tenant', $TenantName)
-        }
-        #endregion
-
+        $driftedData.Add('Parameters', $telemetryDriftedParameters)
+        Add-M365DSCTelemetryEvent -Type 'DriftInfo' -Data $driftedData
         $EventMessage.Append("        </ParametersNotInDesiredState>`r`n") | Out-Null
         $EventMessage.Append("    </ConfigurationDrift>`r`n") | Out-Null
         $EventMessage.Append("    <DesiredValues>`r`n") | Out-Null
@@ -950,6 +984,7 @@ function Test-M365DSCParameterState
                 $Value = "`$null"
             }
             $EventMessage.Append("        <Param Name =`"$key`">$Value</Param>`r`n") | Out-Null
+            $DriftObject.DesiredValues.Add($key, $value)
         }
         $EventMessage.Append("    </DesiredValues>`r`n") | Out-Null
         $EventMessage.Append("    <CurrentValues>`r`n") | Out-Null
@@ -961,10 +996,12 @@ function Test-M365DSCParameterState
                 $Value = "`$null"
             }
             $EventMessage.Append("        <Param Name =`"$key`">$Value</Param>`r`n") | Out-Null
+            $DriftObject.CurrentValues.Add($key, $value)
         }
         $EventMessage.Append("    </CurrentValues>`r`n") | Out-Null
         $EventMessage.Append('</M365DSCEvent>') | Out-Null
 
+        $Global:CCMCurrentDriftInfo = $DriftObject
         Add-M365DSCEvent -Message $EventMessage.ToString() -EventType 'Drift' -EntryType 'Warning' `
             -EventID 1 -Source $Source
     }
@@ -1267,9 +1304,9 @@ function Export-M365DSCConfiguration
 
     $data.Add('Path', [System.String]::IsNullOrEmpty($Path))
     $data.Add('FileName', $null -ne [System.String]::IsNullOrEmpty($FileName))
-    $data.Add('Components', $null -ne $Components)
-    $data.Add('Workloads', $null -ne $Workloads)
-    $data.Add('MaxProcesses', $null -ne $MaxProcesses)
+    $data.Add('Components', $Components)
+    $data.Add('Workloads', $Workloads)
+    $data.Add('MaxProcesses', $MaxProcesses)
     #endregion
 
     if ($null -eq $MaxProcesses)
@@ -1288,20 +1325,12 @@ function Export-M365DSCConfiguration
         Write-Verbose -Message 'No existing connections to Microsoft Graph'
     }
 
-    if (-not [System.String]::IsNullOrEmpty($TenantId) -and -not $data.ContainsKey('Tenant'))
-    {
-        $data.Add('Tenant', $TenantId)
-    }
-    else
-    {
-        if ($Credential -and -not $data.ContainsKey('Tenant'))
-        {
-            $tenant = $Credential.UserName.Split('@')[1]
-            $data.Add('Tenant', $tenant)
-        }
-    }
+    $Tenant = Get-M365DSCTenantNameFromParameterSet -ParameterSet $PSBoundParameters
+    $data.Add('Tenant', $Tenant)
+    $currentExportID = (New-Guid).ToString()
+    $data.Add('M365DSCExportId', $currentExportID)
 
-    Add-M365DSCTelemetryEvent -Type 'Export' -Data $data
+    Add-M365DSCTelemetryEvent -Type 'ExportInitiated' -Data $data
     if ($null -ne $Workloads)
     {
         Write-Output -InputObject "Exporting Microsoft 365 configuration for Workloads: $($Workloads -join ', ')"
@@ -1365,6 +1394,8 @@ function Export-M365DSCConfiguration
     # Clear the exported resource instances' names Global variable
     $Global:M365DSCExportedResourceInstancesNames = $null
     $Global:M365DSCExportInProgress = $false
+
+    Add-M365DSCTelemetryEvent -Type 'ExportCompleted' -Data $data
 }
 
 $Script:M365DSCDependenciesValidated = $false
@@ -1381,7 +1412,7 @@ function Confirm-M365DSCDependencies
     [CmdletBinding()]
     param()
 
-    if (-not $Script:M365DSCDependenciesValidated)
+    if (-not $Script:M365DSCDependenciesValidated -and ($null -eq $Global:M365DSCSkipDependenciesValidation -or -not $Global:M365DSCSkipDependenciesValidation))
     {
         Write-Verbose -Message 'Dependencies were not already validated.'
 
@@ -1679,7 +1710,7 @@ function New-M365DSCConnection
         [System.Boolean]
         $SkipModuleReload = $false
     )
-
+    $Global:MaximumFunctionCount = 32767
     if ($Workload -eq 'MicrosoftTeams')
     {
         try
@@ -2949,7 +2980,7 @@ function Uninstall-M365DSCOutdatedDependencies
                 Write-Information -MessageData "Uninstalling $($module.Name) Version {$($module.Version)}"
                 if (Test-Path -Path $($module.Path))
                 {
-                    Remove-Item $($module.Path) -Force -Recurse
+                    Remove-Item $($module.ModuleBase) -Force -Recurse
                 }
             }
             catch
@@ -3330,38 +3361,47 @@ function Get-M365DSCExportContentForResource
     $Results = Format-M365DSCString -Properties $Results `
         -ResourceName $ResourceName
 
+    if ($Script:AllM365DscResources.Count -eq 0)
+    {
+        $Script:AllM365DscResources = Get-DscResource -Module 'Microsoft365Dsc'
+    }
+
     $primaryKey = ''
-    if ($Results.ContainsKey('IsSingleInstance'))
+    $Resource = $Script:AllM365DscResources.Where({ $_.Name -eq $ResourceName })
+    $Keys = $Resource.Properties.Where({ $_.IsMandatory }) | `
+        Select-Object -ExpandProperty Name
+    if ($Keys.Contains('IsSingleInstance'))
     {
         $primaryKey = ''
     }
-    elseif ($Results.ContainsKey('DisplayName'))
+    elseif ($Keys.Contains('DisplayName'))
     {
         $primaryKey = $Results.DisplayName
     }
-    elseif ($Results.ContainsKey('Identity'))
-    {
-        $primaryKey = $Results.Identity
-    }
-    elseif ($Results.ContainsKey('Id'))
-    {
-        $primaryKey = $Results.Id
-    }
-    elseif ($Results.ContainsKey('Name'))
+    elseif ($Keys.Contains('Name'))
     {
         $primaryKey = $Results.Name
     }
-    elseif ($Results.ContainsKey('Title'))
+    elseif ($Keys.Contains('Title'))
     {
         $primaryKey = $Results.Title
     }
-    elseif ($Results.ContainsKey('CdnType'))
+    elseif ($Keys.Contains('Identity'))
     {
-        $primaryKey = $Results.CdnType
+        $primaryKey = $Results.Identity
     }
-    elseif ($Results.ContainsKey('Usage'))
+    elseif ($Keys.Contains('Id'))
     {
-        $primaryKey = $Results.Usage
+        $primaryKey = $Results.Id
+    }
+
+    if ([String]::IsNullOrEmpty($primaryKey) -and `
+        -not $Keys.Contains('IsSingleInstance'))
+    {
+        foreach ($Key in $Keys)
+        {
+            $primaryKey += $Results.$Key
+        }
     }
 
     $instanceName = $ResourceName
@@ -4560,10 +4600,6 @@ function Sync-M365DSCParameter
     if ($PSBoundParameters.ContainsKey('Parameters'))
     {
         $Parameters
-    }
-    else
-    {
-        $global:ALBoundParameters = $Parameters
     }
 }